+++ src/devices.js
@@ -1,7 +1,7 @@
 // Device support loader
 // Part of homebridge-nest-accfactory
 //
-// Code version 2025.07.25
+// Code version 2026.02.14
 // Mark Hulskamp
 'use strict';
 
@@ -21,7 +21,7 @@ async function loadDeviceModules(log, pluginDir = '') {
   let deviceMap = new Map();
   let files = (await fs.readdir(baseDir)).sort();
 
-  log?.debug?.('Using base module v%s', HomeKitDevice.VERSION);
+  log?.debug?.('Base module "v%s"', HomeKitDevice.VERSION);
 
   for (const file of files) {
     if (file.endsWith('.js') === false) {
@@ -57,7 +57,7 @@ async function loadDeviceModules(log, pluginDir = '') {
         }
 
         deviceMap.set(chosenClass.TYPE, entry);
-        log?.info?.('Loaded %s module v%s', chosenClass.TYPE, chosenClass.VERSION);
+        log?.info?.('Loaded %s module "v%s"', chosenClass.TYPE, chosenClass.VERSION);
       }
     } catch (error) {
       log?.warn?.('Failed to load device support file "%s": %s', file, error.message);